multi-services: Add com.endlessm.EknServicesMultiplexer flatpak
This actually adds two build manifests. The first,
com.endlessm.BaseEknServicesMultiplexer is a "base" flatpak which merely
adds the extension points for the extensions defined in each
eos-knowledge-services version (eg, com.endlessm.EknSerivces.Extension).

That flatpak is then used as a "base" for com.endlessm.EknServicesMultiplexer
which will get those extensions mounted at build-time so that we can copy
their contents into the right place in com.endlessm.EknServicesMultiplexer.

com.endlessm.EknServicesMultiplexer itself is made up of a few things. First,
it mounts every available SDK as an extension using the flatpak "tagged extension
names" feature introduced in flatpak 0.11.4. Second, it exports the d-bus service
files for each eos-knowledge-services version, all of which point to a "multiplexer"
binary.

The "multiplexer" binary (eks-multi-search-provider-dispatcher) takes a
single argument, --services-version and uses that to try and find the
highest-versioned compatible SDK and sets LD_LIBRARY_PATH and friends,
then executes the correct eks-search-provider binary that was copied in place
at build time.

https://phabricator.endlessm.com/T21822